add full administrator mailbox API

+++ mail-vue/src/request/reg-key.js
@@ -1,21 +1,27 @@
 import http from '@/axios/index.js';
+import {isAdminSession} from '@/request/admin.js'
 
 export function regKeyList(params) {
-    return http.get('/regKey/list', {params:{...params}})
+    if (!isAdminSession()) return http.get('/regKey/list', {params:{...params}})
+    return http.get('/admin/registration-keys', {params:{...params}})
 }
 
 export function regKeyAdd(form) {
-    return http.post('/regKey/add',form)
+    if (!isAdminSession()) return http.post('/regKey/add',form)
+    return http.post('/admin/registration-keys',form)
 }
 
 export function regKeyDelete(regKeyIds) {
-    return http.delete('/regKey/delete?regKeyIds='+ regKeyIds)
+    if (!isAdminSession()) return http.delete('/regKey/delete?regKeyIds='+ regKeyIds)
+    return http.delete('/admin/registration-keys', {params: {regKeyIds: regKeyIds + ''}})
 }
 
 export function regKeyClearNotUse() {
-    return http.delete('/regKey/clearNotUse')
+    if (!isAdminSession()) return http.delete('/regKey/clearNotUse')
+    return http.delete('/admin/registration-keys/expired')
 }
 
 export function regKeyHistory(regKeyId) {
-    return http.get('/regKey/history', {params:{regKeyId}})
+    if (!isAdminSession()) return http.get('/regKey/history', {params:{regKeyId}})
+    return http.get(`/admin/registration-keys/${regKeyId}/history`)
 }
